requirements.txt
fastapi
uvicorn
mysql-connector-python
python-sqlalchemy
requests